refactor: topic-split combat test support god scenario modules (#4196)
Slice B splits four >300-line scenario tables along existing private-list
seams into topic files with thin aggregators; preserves all public factory
names and scenario metadata.

Refs #4196
